backend/requirements.txt
instructor
openai
fastapi
uvicorn
loguru==0.7.2
#pymgclient==1.3.1
requests==2.31.0
python-dotenv==1.0.1
python-jose
python-multipart
passlib
requests
websockets
cachetools==5.5.0
bcrypt==4.0.1
celery==5.4.0
redis==5.0.8
flower==2.0.1
PyYAML==6.0.2
Jinja2==3.1.4
prometheus-fastapi-instrumentator==7.0.0
neo4j==5.24.0
PyMySQL==1.1.1
alembic==1.13.2
wtforms==3.1.2